requirements-dev.txt
scikit-learn
fastapi
python-dotenv
pandas
pytest
httpx
pytest-asyncio